+++ src/data/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name    = "enso-data"
-version = "0.1.2"
+version = "0.1.3"
 authors = ["Enso Team <contact@luna-lang.org>"]
 edition = "2018"
 
@@ -19,6 +19,13 @@ publish = true
 crate-type = ["rlib", "cdylib"]
 
 [dependencies]
-enso-prelude = { version = "0.1.1" , path = "../prelude" }
-
+enso-prelude = { version = "0.1.8" , path = "../prelude" }
 serde        = { version = "1.0"  , features = ["derive"] }
+
+[dev-dependencies]
+itertools = "0.9.0"
+criterion = "0.3"
+
+[[bench]]
+name = "bench_tree_query"
+harness = false
