+++ src/prelude/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name    = "enso-prelude"
-version = "0.1.9"
+version = "0.1.8"
 authors = ["Enso Team <enso-dev@enso.org>"]
 edition = "2018"
 
@@ -84,8 +84,3 @@ features = [
 
 [dev-dependencies]
 wasm-bindgen-test = "0.2"
-criterion         = "0.3"
-
-[[bench]]
-name = "bench_tree_query"
-harness = false
