+++ src/prelude/src/data.rs
@@ -1,13 +1,11 @@
 //! Generic data types and utilities.
 
 pub mod at_least_one_of_two;
-pub mod hash_tree;
 pub mod monoid;
 pub mod non_empty_vec;
 pub mod semigroup;
 
 pub use at_least_one_of_two::*;
-pub use hash_tree::*;
 pub use monoid::*;
 pub use non_empty_vec::*;
 pub use semigroup::*;
